+++ internal/engine/internal/sessionresolver/resolvermaker.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/apex/log"
+	"github.com/ooni/probe-cli/v3/internal/bytecounter"
 	"github.com/ooni/probe-cli/v3/internal/engine/netx"
-	"github.com/ooni/probe-cli/v3/internal/engine/netx/bytecounter"
 )
 
 // resolvemaker contains rules for making a resolver.
