+++ internal/engine/netx/dialer/bytecounter.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"net"
 
-	"github.com/ooni/probe-cli/v3/internal/engine/netx/bytecounter"
+	"github.com/ooni/probe-cli/v3/internal/bytecounter"
 )
 
 // byteCounterDialer is a byte-counting-aware dialer. To perform byte counting, you
